wip #359 move checkout thank you to checkut blueprint, out of views.py
