+++ subscribie/blueprints/checkout/templates/thankyou.html
@@ -1,0 +1,21 @@
+{% extends "layout.html" %}
+{% block title %}Thank you!{% endblock %}
+
+{% block hero %}
+
+  <div class="container">
+
+    <svg style="max-width: 350px; width:75%;" viewBox="0 0 173 107" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
+        <title>Subscription Builder - Success</title>
+        <defs></defs>
+        <g id="Subscription-Builder---Success" stroke="none" stroke-width="1" fill="none" fill-rule="evenodd" stroke-linecap="round">
+            <path class="tick" d="M65,53.9213658 L78.2769791,67.4578625 C79.050432,68.2464336 80.3167026,68.2586894 81.1052737,67.4852365 C81.1183336,67.472427 81.1312171,67.459439 81.1439206,67.446276 L108.597255,39" id="Line" stroke="#FFFFFF" stroke-width="10"></path>
+        </g>
+    </svg>
+
+    <h1 style="color:white" class="title is-1 is-size-3-mobile">Order Complete!</h1>
+    <h2 style="color:white" class="subtitle is-4 is-size-5-mobile">We've received your order and will be in contact soon.</h2>
+
+  </div>
+
+{% endblock %}
